backend/requirements.txt
google-generativeai
python-dotenv
cassandra-driver
Flask
flask-cors